+++ app/javascript/components/PlayerActionsModal/PlayerActionsModal.jsx
@@ -15,10 +15,10 @@ export const PlayerActionsModal = ({ lineupPlayer, onMakeCaptain, onClose }) => 
       <h2 className="pr-8">{localizeValue(lineupPlayer.player.name)}</h2>
       {onMakeCaptain ? (
         <div className="flex flex-col items-start">
-          <button className="btn-primary mb-2" onClick={() => onMakeCaptain(lineupPlayer.uuid, 'captain')}>
+          <button className="btn-primary mb-2" onClick={() => onMakeCaptain(lineupPlayer, 'captain')}>
             {strings.playerActions.makeCaptain}
           </button>
-          <button className="btn-primary" onClick={() => onMakeCaptain(lineupPlayer.uuid, 'assistant')}>
+          <button className="btn-primary" onClick={() => onMakeCaptain(lineupPlayer, 'assistant')}>
             {strings.playerActions.makeAssistant}
           </button>
         </div>